Update to work with proxy_pass and static files
- Adds a new volume to mount the staticfiles to nginx
- Updates nginx to:
  - Use the webapp service port
  - Reference the staticfiles location

Testing:

You should be able to bring up the stack and load the application at http://localhost.

This is (nginx -> webapp:8000) and serving static files from nginx.

NOTE!!! In my test on another branch I had to merge in the latest from `development`. I was going to use my test branch as the base for a PR to this one but it contained 900+ changes from the merge and I didn't want to make your head explode so just decided to modify this existing draft PR.