Updates the Nginx location block to work properly when we set `MAIN_PAGE` in the production context to serve at `/signals/static/...`.
This uses a regex to match on anyting that comes before `/static`, and then capture anything after into `$1` so that we can pass that as a file path to properly load the static files.

Previously this only worked when we were server at the bare URL, but now it works when we are serving at `/signals/` (or potentially other URL "prefixes") as well.